requirements.txt
asgiref==3.3.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
django-model-utils==4.0.0
django-random-queryset==0.1.3
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
